Moved converted files to a src/ folder.
